third_party/blink/web_tests/external/wpt/webrtc/RTCPeerConnection-createOffer-expected.txt

This is a testharness.js-based test.
[FAIL] When media stream is added when createOffer() is running in parallel, the result offer should contain the new media stream
  assert_equals: Expect m=audio line to be found in offer SDP expected 1 but got 0
[FAIL] createOffer() should fail when signaling state is not stable or have-local-offer
  assert_unreached: Should have rejected: undefined Reached unreachable code
Harness: the test ran to completion.
